A fix for spray/spray#208
The default RejectionHandler attempts to construct a response to a rejected
request using the Spray marshalling framework. Unfortunately, if the original
request does not have an Accept header that supports 'text/plain', the
marshalling framework generates a rejection of it's own, and it's apparently
turtles all the way down. This fix just causes the default RejectionHandler
to construct raw HttpResponse objects on it's own and complete the request
directly.